Share the integration-test helpers instead of copying them
`assert_close` and `collect_gguf` each had two copies across the
integration suites. Duplication matters more in a test than in library
code: a tolerance or a comparison that drifts between copies makes one
suite quietly weaker than the other, and a test that has stopped
checking what it claims to check looks exactly like a test that passes.

Rust compiles every file under `tests/` as its own crate, so a helper
only one suite uses is dead code in the others -- `#[allow(dead_code)]`
sits on the module rather than each item, since which suite uses what
is not a property worth maintaining.

That takes the workspace's cross-file duplicate function bodies from 14
to 1, and the one left is deliberate: `print_available_devices`, whose
merge would cost ferrox-core two required backend dependencies where
today they are optional.

Workspace tests, clippy and fmt clean.
